got it working
