+++ deploy/helm/groundlight-edge-endpoint/templates/service-account.yaml
@@ -44,6 +44,9 @@ rules:
   resources: ["secrets"]
   # Needed to refresh ECR credentials
   verbs: ["create", "delete", "get", "patch"]
+- apiGroups: ["metrics.k8s.io"]
+  resources: ["pods"]
+  verbs: ["list"]
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: RoleBinding
